alotech/requirements.txt
Django==3.1.6
django-redis==4.12.1
